Move reader module to its own file
That way the structure is a bit more clear.
