+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod config;
+pub mod reader;
